refactor(Coin) : Coin 도메인 위치 변경
