+++ src/main/java/vendingmachine/domain/Machine.java
@@ -1,7 +1,5 @@
 package vendingmachine.domain;
 
-import vendingmachine.Coin;
-
 public class Machine {
 	private final int money;
 	private Coins coins;
